requirements.txt
streamlit
diffusers
torch
transformers
git+https://github.com/facebookresearch/segment-anything.git
pascal_voc_writer
roboflow 
dataclasses-json 
supervision
torchvision
google-generativeai